packages/llm-clients/pyproject.toml
[project]
name = "aurixa-llm-clients"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "openai>=1.60.0",
    "anthropic>=0.43.0",
    "google-generativeai>=0.8.0",
    "tiktoken>=0.8.0",
    "loguru>=0.7.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"

[tool.hatch.build.targets.wheel]
packages = ["src/aurixa_llm"]
